Changes samuel
