+++ apps/web-app/components/CalendarSessionModal/Step2.tsx
@@ -1,29 +1,16 @@
-import { ComponentType, useEffect, useRef, useState } from "react"
-import dynamic, { Loader } from "next/dynamic"
-import DatePicker from "react-datepicker"
-import axios from "axios"
 import moment from "moment"
-import { ToastContainer, toast } from "react-toastify"
 import NextImage from "next/image"
 import { IoMdArrowBack } from "react-icons/io"
-import { EditorState } from "draft-js"
-import { stateToHTML } from "draft-js-export-html"
-import MaskedInput from "react-text-mask"
-import { EditorProps } from "react-draft-wysiwyg"
-import { stateFromHTML } from "draft-js-import-html"
-import TimeDropdown from "../TimeDropdown"
+import { Parser } from "html-to-react"
 
-import { TracksDTO, FormatDTO, LevelDTO, LocationDTO, EventTypeDTO, EventsDTO, SessionsDTO } from "../../types"
-import SlotsAvailableModal from "../SlotsAvailableModal"
+import Loading from "../Loading"
 
 type NewSessionState = {
     description: string
     equipment: string
     event_id: number
-    event_item_id: number
-    event_slug: string
     event_type: string
-    duration: string
+    maxRsvp: string
     format: string
     hasTicket: boolean
     info: string
@@ -31,594 +18,63 @@ type NewSessionState = {
     location: string
     custom_location: string
     name: string
-    startDate: Date
-    startTime: string
+    startDate: string
     endTime: string
+    startTime: string
     tags: string[]
     team_members: {
         name: string
         role: string
     }[]
     track: string
+    event_slug: string
+    event_item_id: number
 }
 
 type Props = {
+    setSteps: (step: number) => void
     newSession: NewSessionState
-    setNewSession: (newEvent: NewSessionState) => void
-    setSteps: (steps: number) => void
-    sessions: SessionsDTO[]
-    events: EventsDTO[]
+    handleSubmit: Function
+    isLoading: boolean
+    amountTickets: string
 }
 
-// @ts-ignore
-const loadEditor: Loader<EditorProps> = async () => {
-    const mod = await import("react-draft-wysiwyg")
-    return mod.Editor as ComponentType<EditorProps>
-}
-
-const Editor = dynamic<EditorProps>(loadEditor, { ssr: false })
-
-const Step2 = ({ events, newSession, setNewSession, setSteps, sessions }: Props) => {
-    const { name, team_members, startDate, tags, startTime, duration, location, event_id, description } = newSession
-    const wraperRef = useRef(null)
-    const [teamMember, setTeamMember] = useState({ name: "", role: "Speaker" })
-    const [tag, setTag] = useState("")
-    const [rerender, setRerender] = useState(true)
-    // const [suggestions, setSuggestions] = useState<UserDTO[]>([])
-    const [display, setDisplay] = useState(false)
-    const [tracksOpt, setTracksOpt] = useState<TracksDTO[]>()
-    const [formatsOpt, setFormatsOpt] = useState<FormatDTO[]>()
-    const [levelsOpt, setLevelsOpt] = useState<LevelDTO[]>()
-    const [locationsOpt, setLocationsOpt] = useState<LocationDTO[]>()
-    const [eventTypesOpt, setEventTypesOpt] = useState<EventTypeDTO[]>()
-    const [openSlotsModal, setOpenSlotsModal] = useState(false)
-    const [filteredSessionsModal, setFilteredSessionsModal] = useState<SessionsDTO[]>([])
-
-    const htmlToEditorState = (html: string) => {
-        const contentState = stateFromHTML(html)
-        return EditorState.createWithContent(contentState)
-    }
-
-    const [richTextEditor, setRichTextEditor] = useState<EditorState>(htmlToEditorState(description))
-
-    const onEditorStateChange = (editorState: EditorState) => {
-        setRichTextEditor(editorState)
-        const html = stateToHTML(editorState.getCurrentContent())
-
-        setNewSession({ ...newSession, description: html })
-    }
-
-    const handleSelectEventTrack = (index: number) => {
-        const selectedEvent = events[index]
-        if (selectedEvent.id === 101) {
-            setNewSession({
-                ...newSession,
-                location: "Other",
-                event_id: selectedEvent.id,
-                event_slug: selectedEvent.slug,
-                event_item_id: selectedEvent.item_id
-            })
-        } else {
-            setNewSession({
-                ...newSession,
-                location: "",
-                event_id: selectedEvent.id,
-                event_slug: selectedEvent.slug,
-                event_item_id: selectedEvent.item_id
-            })
-        }
-    }
-
-    const handleAddTeamMember = () => {
-        setNewSession({ ...newSession, team_members: [...newSession.team_members, teamMember] })
-        setTeamMember({ name: "", role: "Speaker" })
-        setDisplay(false)
-
-        const selectedRole = teamMember.role
-        setTeamMember({ name: "", role: selectedRole })
-    }
-
-    const handleRemoveTeamMember = (index: number) => {
-        team_members.splice(index, 1)
-        setRerender(!rerender)
-    }
-
-    const handleAddTag = () => {
-        setNewSession({ ...newSession, tags: [...newSession.tags, tag] })
-        setTag("")
-    }
-
-    const handleRemoveTag = (e: any, index: number) => {
-        e.preventDefault()
-        tags.splice(index, 1)
-        setRerender(!rerender)
-    }
-
-    const handleClickOutside = (event: MouseEvent) => {
-        const { current: wrap } = wraperRef as { current: HTMLElement | null }
-
-        if (wrap && !wrap.contains(event.target as Node)) {
-            setDisplay(false)
-        }
-    }
-
-    const fetchTraks = async () => {
-        await axios
-            .get("/api/fetchTracks")
-            .then((res) => {
-                setTracksOpt(res.data)
-            })
-            .catch((err) => console.log(err))
-    }
-
-    const fetchLevels = async () => {
-        await axios
-            .get("/api/fetchLevels")
-            .then((res) => {
-                setLevelsOpt(res.data)
-            })
-            .catch((err) => console.log(err))
-    }
-
-    const fetchEventTypes = async () => {
-        await axios
-            .get("/api/fetchEventTypes")
-            .then((res) => {
-                setEventTypesOpt(res.data)
-            })
-            .catch((err) => console.log(err))
-    }
-
-    const fetchFormats = async () => {
-        await axios
-            .get("/api/fetchFormats")
-            .then((res) => {
-                setFormatsOpt(res.data)
-            })
-            .catch((err) => console.log(err))
-    }
-
-    const fetchLocations = async () => {
-        await axios
-            .get("/api/fetchLocations")
-            .then((res) => {
-                setLocationsOpt(res.data)
-            })
-            .catch((err) => console.log(err))
-    }
-
-    useEffect(() => {
-        Promise.all([fetchLevels(), fetchEventTypes(), fetchFormats(), fetchLocations(), fetchTraks()])
-    }, [])
-
-    useEffect(() => {
-        document.addEventListener("mousedown", handleClickOutside)
-
-        return () => {
-            document.removeEventListener("mousedown", handleClickOutside)
-        }
-    }, [])
-
-    const isOverlapping = ({ filteredSeshs }: { filteredSeshs: SessionsDTO[] }) => {
-        const formatDate = moment.utc(newSession.startDate).format("YYYY-MM-DD")
-        const newSessionStart = moment.utc(`${formatDate}T${newSession.startTime}`)
-        const newSessionEnd = newSessionStart.clone().add(newSession.duration, "minutes")
-
-        for (const idx of filteredSeshs) {
-            const sessionStart = moment.utc(`${idx.startDate}T${idx.startTime}`)
-            const sessionEnd = sessionStart.clone().add(idx.duration, "minutes")
-
-            if (
-                (newSessionStart.isSameOrAfter(sessionStart) && newSessionStart.isBefore(sessionEnd)) ||
-                (newSessionEnd.isAfter(sessionStart) && newSessionEnd.isSameOrBefore(sessionEnd)) ||
-                (newSessionStart.isSameOrBefore(sessionStart) && newSessionEnd.isSameOrAfter(sessionEnd))
-            ) {
-                return true
-            }
-        }
-
-        return false
-    }
-
-    const handleNextStep = () => {
-        if (
-            newSession.name.length === 0 ||
-            newSession.description.length === 0 ||
-            newSession.location === "" ||
-            newSession.team_members.length === 0 ||
-            newSession.startTime === ""
-        ) {
-            return toast.error("Please fill all inputs required.", {
-                position: "top-center",
-                autoClose: 3000,
-                hideProgressBar: false,
-                closeOnClick: true,
-                pauseOnHover: true,
-                draggable: true,
-                progress: undefined,
-                theme: "light"
-            })
-        }
-
-        if (newSession.duration === "0") {
-            return toast.error("Please fill duration time field.", {
-                position: "top-center",
-                autoClose: 3000,
-                hideProgressBar: false,
-                closeOnClick: true,
-                pauseOnHover: true,
-                draggable: true,
-                progress: undefined,
-                theme: "light"
-            })
-        }
-
-        const selectedLocation = newSession.location.toLocaleLowerCase()
-
-        const filteredSeshs = sessions
-            .filter((item) => item.location.toLocaleLowerCase() === selectedLocation)
-            .filter((item) => {
-                const formatDate = moment.utc(newSession.startDate).format("YYYY-MM-DD")
-
-                const selectedDate = moment.utc(formatDate)
-                const newSessionStartDate = moment.utc(item.startDate)
-
-                return selectedDate.isSame(newSessionStartDate)
-            })
-
-        if (isOverlapping({ filteredSeshs })) {
-            return toast.error("Session already booked on that Date and Time.", {
-                position: "top-center",
-                autoClose: 3000,
-                hideProgressBar: false,
-                closeOnClick: true,
-                pauseOnHover: true,
-                draggable: true,
-                progress: undefined,
-                theme: "light"
-            })
-        }
-        setSteps(3)
-    }
-
-    const handleOpenSlotsAvailable = () => {
-        const selectedLocation = newSession.location.toLocaleLowerCase()
-
-        const filteredSeshs = sessions
-            .filter((item) => item.location.toLocaleLowerCase() === selectedLocation)
-            .filter((item) => {
-                const formatDate = moment.utc(newSession.startDate).format("YYYY-MM-DD")
-
-                const selectedDate = moment.utc(formatDate)
-                const newSessionStartDate = moment.utc(item.startDate)
-
-                return selectedDate.isSame(newSessionStartDate)
-            })
-
-        setFilteredSessionsModal(filteredSeshs)
-    }
+const Step2 = ({ setSteps, newSession, handleSubmit, isLoading }: Props) => {
+    const parser = new Parser()
+    const reactContent = parser.parse(newSession.description)
 
     return (
-        <div className="flex flex-col w-full">
-            <SlotsAvailableModal
-                closeModal={setOpenSlotsModal}
-                isOpen={openSlotsModal}
-                sessions={filteredSessionsModal}
-                startDate={startDate}
-                location={location}
-            />
-            <ToastContainer
-                position="top-center"
-                autoClose={3000}
-                hideProgressBar={false}
-                newestOnTop={false}
-                closeOnClick
-                rtl={false}
-                pauseOnFocusLoss
-                draggable
-                pauseOnHover
-                theme="light"
-            />
-            <div className="flex flex-col gap-1 my-2 w-full">
-                <label htmlFor="name" className="font-[600]">
-                    Session Title*
-                </label>
-                <input
-                    className="border-[#C3D0CF] border-2 p-1 rounded-[8px] h-[42px]"
-                    type="text"
-                    id="name"
-                    placeholder="Event name"
-                    value={name}
-                    onChange={(e) => setNewSession({ ...newSession, name: e.target.value })}
-                />
-            </div>
-
-            <div className="flex flex-col gap-1 my-2 w-full">
-                <label htmlFor="name" className="font-[600]">
-                    Event Track*
-                </label>
-                <select
-                    id="eventTrack"
-                    name="eventTrack"
-                    className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px] w-full"
-                    onChange={(e) => handleSelectEventTrack(parseInt(e.target.value))}
-                >
-                    {events.map((item, index) => (
-                        <option key={item.id} value={index}>
-                            {item.name}
-                        </option>
-                    ))}
-                </select>
-            </div>
-
-            {event_id !== 101 && (
-                <div className="flex flex-col gap-1 w-full my-2">
-                    <label htmlFor="location" className="font-[600]">
-                        Location*
-                    </label>
-                    <select
-                        id="location"
-                        name="location"
-                        value={newSession.location}
-                        className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px] w-full"
-                        onChange={(e) => setNewSession({ ...newSession, location: e.target.value })}
-                    >
-                        <option value="Select Location">Select Location</option>
-                        {locationsOpt &&
-                            locationsOpt.map((item, index) => (
-                                <option key={index} value={item.location}>
-                                    {item.location}
-                                </option>
-                            ))}
-                    </select>
+        <div className="flex flex-col w-full gap-8 bg-white rounded-lg mt-5">
+            <div className="flex flex-col gap-4">
+                <div className="flex flex-col">
+                    <h1 className="text-[16px] font-[600]">{newSession.track}</h1>
+                    <h1 className="text-[22px] font-[600]">{newSession.name}</h1>
                 </div>
-            )}
-
-            <div className="flex flex-row w-full gap-5 my-2">
-                <div className="flex flex-col w-2/6">
-                    <label className="font-[600]">Date*</label>
-                    <DatePicker
-                        className="border-[#C3D0CF] border-2 p-1 rounded-[8px] h-[42px] w-full"
-                        selected={startDate}
-                        onChange={(e) => setNewSession({ ...newSession, startDate: e as Date })}
-                        minDate={moment().toDate()}
-                    />
+                <div className="flex items-center gap-2">
+                    <NextImage src="/vector-calendar.svg" width={20} height={20} />
+                    <h1>{moment.utc(newSession.startDate).format("dddd, MMMM DD")}</h1>
                 </div>
-                <div className="flex flex-col w-2/6">
-                    <label htmlFor="startTime" className="font-[600]">
-                        Start Time*
-                    </label>
-                    <TimeDropdown
-                        id="startTime"
-                        value={newSession.startTime}
-                        onChange={(e: any) => setNewSession({ ...newSession, startTime: e.target.value })}
-                    />
+                <div className="flex items-center gap-2">
+                    <NextImage src="/vector-location.svg" width={20} height={20} />
+                    <h1 className="text-[18px]">
+                        {newSession.location === "Other" ? newSession.custom_location : newSession.location}
+                    </h1>
                 </div>
-                <div className="flex flex-col w-2/6">
-                    <label htmlFor="startTime" className="font-[600]">
-                        End Time*
-                    </label>
-                    <TimeDropdown
-                        id="endTime"
-                        value={newSession.endTime}
-                        onChange={(e: any) => setNewSession({ ...newSession, endTime: e.target.value })}
-                    />
+                <div className="flex items-center gap-2">
+                    <NextImage src="/vector-clock.svg" width={20} height={20} />
+                    <h1>
+                        {newSession.startTime}-{newSession.endTime}
+                    </h1>
+                </div>
+                <div className="flex items-center gap-2">
+                    <NextImage src="/vector-clock.svg" width={20} height={20} />
+                    <h1>{newSession.maxRsvp} RSVPs Allowed</h1>
+                </div>
+                <div className="flex items-center gap-2 mt-10">
+                    <h1 className="text-[18px]">{reactContent}</h1>
                 </div>
             </div>
-
-            {location === "Other" ? (
-                <div className="flex flex-col gap-1 w-full mt-2">
-                    <label htmlFor="custom_location" className="font-[600]">
-                        Specify location
-                    </label>
-                    <input
-                        type="text"
-                        placeholder="Specify Location"
-                        className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px] w-full"
-                        value={newSession.custom_location}
-                        onChange={(e) => setNewSession({ ...newSession, custom_location: e.target.value })}
-                    />
-                </div>
-            ) : (
-                ""
-            )}
-
-            <div className="flex flex-col gap-1 my-2 w-full">
-                <label htmlFor="info" className="font-[600]">
-                    Description*
-                </label>
-                <div className="w-full h-[400px] p-4 border border-gray-300 rounded overflow-scroll">
-                    {richTextEditor && (
-                        // @ts-ignore
-                        <Editor
-                            editorState={richTextEditor}
-                            onEditorStateChange={onEditorStateChange}
-                            wrapperClassName="wrapper-class"
-                            editorClassName="editor-class"
-                            toolbarClassName="toolbar-class"
-                        />
-                    )}
-                </div>
-            </div>
-
-            {startDate && location && location !== "Other" && event_id !== 101 && (
-                <div className="flex flex-col justify-start my-2">
-                    <button
-                        onClick={() => {
-                            handleOpenSlotsAvailable()
-                            setOpenSlotsModal(true)
-                        }}
-                        className="flex flex-row font-[600] w-full justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
-                    >
-                        Check Unavailable Slots ({moment.utc(startDate).format("MM/DD/YYYY")})
-                    </button>
-                </div>
-            )}
-
-            <div className="flex flex-col gap-4 w-full my-8">
-                <div className="flex flex-col md:flex-row w-full gap-4">
-                    <div className="flex flex-col md:w-3/6 w-full">
-                        <label htmlFor="team-members" className="font-[600]">
-                            Organizers*
-                        </label>
-                        <input
-                            id="tags"
-                            type="text"
-                            className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px] w-full"
-                            placeholder="Add team member"
-                            value={teamMember.name}
-                            onChange={(e) => setTeamMember({ ...teamMember, name: e.target.value })}
-                        />
-                    </div>
-                    <div className="flex flex-col w-full md:w-3/6">
-                        <label htmlFor="role" className="font-[600]">
-                            Role*
-                        </label>
-                        <select
-                            id="role"
-                            name="role"
-                            className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px] w-full"
-                            onChange={(e) => setTeamMember({ ...teamMember, role: e.target.value })}
-                        >
-                            <option value="Speaker">Speaker</option>
-                            <option value="Organizer">Organizer</option>
-                            <option value="Facilitator">Facilitator</option>
-                        </select>
-                    </div>
-                </div>
-                <div className="flex flex-col gap-4">
-                    <button
-                        className="flex flex-row font-[600] w-full justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
-                        onClick={() => handleAddTeamMember()}
-                    >
-                        ADD
-                    </button>
-                    <ul className="flex flex-row items-center">
-                        {team_members.map((item, index) => (
-                            <div
-                                className="flex flex-row items-center bg-[#E4EAEA] py-[4px] px-[8px] gap-[8px] text-sm rounded-[4px] mr-[8px] cursor-pointer"
-                                key={index}
-                                onClick={() => handleRemoveTeamMember(index)}
-                            >
-                                {item.role === "Speaker" && (
-                                    <NextImage src={"/user-icon-6.svg"} alt="user-icon-6" width={24} height={24} />
-                                )}
-                                {item.role === "Organizer" && (
-                                    <NextImage src={"/user-icon-4.svg"} alt="user-icon-6" width={24} height={24} />
-                                )}
-                                {item.role === "Facilitator" && (
-                                    <NextImage src={"/user-icon-5.svg"} alt="user-icon-6" width={24} height={24} />
-                                )}
-                                <p className="text-[#1C2928] font-[400] text-[16px]">
-                                    {item.role}: <span className="font-[600] capitalize">{item.name}</span>
-                                </p>
-                            </div>
-                        ))}
-                    </ul>
-                </div>
-            </div>
-            <div className="flex flex-col gap-4 w-full mb-8">
-                <div className="flex flex-col gap-4">
-                    <label htmlFor="tags" className="font-[600]">
-                        Tags
-                    </label>
-                    <input
-                        id="tags"
-                        type="text"
-                        className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px] w-full"
-                        placeholder="add tag"
-                        value={tag}
-                        onChange={(e) => setTag(e.target.value)}
-                    />
-
-                    <button
-                        className="flex flex-row font-[600] justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
-                        onClick={() => handleAddTag()}
-                    >
-                        ADD
-                    </button>
-                </div>
-                <ul className="flex flex-row items-center">
-                    {newSession.tags.map((item, index) => (
-                        <div
-                            key={index}
-                            className="bg-[#E4EAEA] py-[4px] px-[8px] text-sm rounded-[4px] cursor-pointe mr-[8px] cursor-pointer"
-                            onClick={(e) => handleRemoveTag(e, index)}
-                        >
-                            {item}
-                        </div>
-                    ))}
-                </ul>
-            </div>
-            <div className="flex flex-col gap-1 my-2">
-                <label htmlFor="tags">Track:</label>
-                <select
-                    id="track"
-                    name="track"
-                    className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px]"
-                    onChange={(e) => setNewSession({ ...newSession, track: e.target.value })}
-                    value={newSession.track}
-                >
-                    {tracksOpt &&
-                        tracksOpt.map((item, index) => (
-                            <option key={index} value={item.type}>
-                                {item.type}
-                            </option>
-                        ))}
-                </select>
-            </div>
-            <div className="flex flex-col gap-1 my-2">
-                <label htmlFor="tags">Format:</label>
-                <select
-                    id="format"
-                    name="format"
-                    value={newSession.format}
-                    className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px]"
-                    onChange={(e) => setNewSession({ ...newSession, format: e.target.value })}
-                >
-                    {formatsOpt &&
-                        formatsOpt.map((item, index) => (
-                            <option key={index} value={item.format}>
-                                {item.format}
-                            </option>
-                        ))}
-                </select>
-            </div>
-            <div className="flex flex-col gap-1 my-2">
-                <label htmlFor="type">Type:</label>
-                <select
-                    id="type"
-                    name="type"
-                    value={newSession.event_type}
-                    className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px]"
-                    onChange={(e) => setNewSession({ ...newSession, event_type: e.target.value })}
-                >
-                    {eventTypesOpt &&
-                        eventTypesOpt.map((item, index) => (
-                            <option key={index} value={item.type}>
-                                {item.type}
-                            </option>
-                        ))}
-                </select>
-            </div>
-            <div className="flex flex-col gap-1 my-2">
-                <label htmlFor="level">Experience Level:</label>
-                <select
-                    id="level"
-                    name="level"
-                    value={newSession.level}
-                    className="border-[#C3D0CF] bg-white border-2 p-1 rounded-[8px] h-[42px]"
-                    onChange={(e) => setNewSession({ ...newSession, level: e.target.value })}
-                >
-                    {levelsOpt &&
-                        levelsOpt.map((item, index) => (
-                            <option key={index} value={item.level}>
-                                {item.level}
-                            </option>
-                        ))}
-                </select>
-            </div>
-            <div className="w-full flex flex-col md:flex-row gap-5 justify-center items-center my-5">
+            <div className="w-full flex flex-col md:flex-row gap-5 justify-center items-center mb-10">
                 <button
                     type="button"
                     className="w-full flex flex-row border-zulalu-primary border font-[600] justify-center items-center py-[8px] px-[16px] gap-[8px] bg-white rounded-[8px] text-black text-[16px]"
@@ -629,10 +85,11 @@ const Step2 = ({ events, newSession, setNewSession, setSteps, sessions }: Props)
                 </button>
                 <button
                     type="button"
-                    className="w-full lex flex-row font-[600] justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
-                    onClick={() => handleNextStep()}
+                    disabled={isLoading}
+                    className="w-full flex flex-row font-[600] justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
+                    onClick={() => handleSubmit()}
                 >
-                    NEXT
+                    {isLoading ? <Loading size="xs" /> : "CREATE SESSION!"}
                 </button>
             </div>
         </div>
