+++ apps/web-app/components/EditSessionModal/ModalSteps.tsx
@@ -3,22 +3,9 @@ type Props = {
 }
 
 const ModalSteps = ({ steps }: Props) => (
-    <div className="flex justify-start gap-3 w-full my-10">
-        <div
-            className={`w-[10px] h-[10px] rounded-full ${
-                steps >= 1 ? "bg-zulalu-primary" : "bg-[#D3DDDC]"
-            }`}
-        />
-        <div
-            className={`w-[10px] h-[10px] rounded-full ${
-                steps >= 2 ? "bg-zulalu-primary" : "bg-[#D3DDDC]"
-            }`}
-        />
-        <div
-            className={`w-[10px] h-[10px] rounded-full ${
-                steps >= 3 ? "bg-zulalu-primary" : "bg-[#D3DDDC]"
-            }`}
-        />
+    <div className="flex justify-start gap-3 w-full my-4">
+        <div className={`w-[10px] h-[10px] rounded-full ${steps >= 1 ? "bg-zulalu-primary" : "bg-[#D3DDDC]"}`} />
+        <div className={`w-[10px] h-[10px] rounded-full ${steps >= 2 ? "bg-zulalu-primary" : "bg-[#D3DDDC]"}`} />
     </div>
 )
