+++ apps/web-app/styles/globals.css
@@ -6,6 +6,11 @@
 @tailwind components;
 @tailwind utilities;
 
+.custom-shadow {
+    @apply shadow;
+    box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+}
+
 .pretix-widget-attribution a {
     @apply hidden;
 }
