+++ apps/web-app/templates/CalendarPage/index.tsx
@@ -30,7 +30,6 @@ const CalendarPage = ({ sessions, events }: Props) => {
     // const isOrganizer = userRole === "organizer"
     const isOrganizer = true
 
-
     /* Begin DatePicker code */
     const [openDatePicker, setOpenDatePicker] = useState(false)
     const [datePickerDescription, setDatePickerDescription] = useState("FULL PROGRAM")
@@ -190,20 +189,12 @@ const CalendarPage = ({ sessions, events }: Props) => {
                 </div>
 
                 {userInfo && isOrganizer ? (
-                    <>
-                        <button
-                            className="flex md:hidden w-full flex-row font-[600] justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
-                            onClick={() => setOpenAddSessionModal(true)}
-                        >
-                            CREATE SESSION
-                        </button>
-                        <CalendarSessionModal
-                            closeModal={setOpenAddSessionModal}
-                            isOpen={openAddSessionModal}
-                            events={events}
-                            sessions={sessions}
-                        />
-                    </>
+                    <button
+                        className="flex md:hidden w-full flex-row font-[600] justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
+                        onClick={() => setOpenAddSessionModal(true)}
+                    >
+                        CREATE SESSION
+                    </button>
                 ) : (
                     ""
                 )}
@@ -233,24 +224,23 @@ const CalendarPage = ({ sessions, events }: Props) => {
                             <h1 className="text-[24px] md:text-[40px] text-[#37352F] font-[600]">Sessions</h1>
 
                             {userInfo && isOrganizer ? (
-                                <>
-                                    <button
-                                        className="hidden md:flex flex-row font-[600] justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
-                                        onClick={() => setOpenAddSessionModal(true)}
-                                    >
-                                        CREATE SESSION
-                                    </button>
-                                    <CalendarSessionModal
-                                        closeModal={setOpenAddSessionModal}
-                                        isOpen={openAddSessionModal}
-                                        events={events}
-                                        sessions={sessions}
-                                    />
-                                </>
+                                <button
+                                    className="hidden md:flex flex-row font-[600] justify-center items-center py-[8px] px-[16px] gap-[8px] bg-[#35655F] rounded-[8px] text-white text-[16px]"
+                                    onClick={() => setOpenAddSessionModal(true)}
+                                >
+                                    CREATE SESSION
+                                </button>
                             ) : (
                                 ""
                             )}
                         </div>
+                        <CalendarSessionModal
+                            closeModal={setOpenAddSessionModal}
+                            isOpen={openAddSessionModal}
+                            events={events}
+                            event={null}
+                            sessions={sessions}
+                        />
 
                         <div className="flex flex-col md:flex-row justify-center items-start md:items-start gap-5 w-full md:w-auto">
                             <div className="flex flex-col relative w-full md:w-[150px]" ref={localtionRef}>
